classifier/requirements.txt
numpy
pytorch
yaml
tensorboardX
zipfile
sklearn
torchvision
termcolor
skimage
json
